Add Hibernate UserType abstraction for MicroTypes
To persist or read MicroTypes it is necessary to tell Hibernate how to
handle such values since those will need to be unwrapped when persisting
and wrapped into the MicroType when reading.

The core to the implementation are the classes ImmutableUserType and
MicroTypeUserType. From there the specific types can be built on top.

The first basic type to be supported is String as
StringMicroTypeUserType.